lstm move
