+++ python/ray/rllib/policy/rnn_sequencing.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-"""LSTM support for RLlib.
+"""RNN utils for RLlib.
 
 The main trick here is that we add the time dimension at the last moment.
 The non-LSTM layers of the model see their inputs as one flat batch. Before
@@ -18,6 +18,37 @@ current algorithms: https://github.com/ray-project/ray/issues/2992
 import numpy as np
 
 from ray.rllib.utils.annotations import DeveloperAPI
+from ray.rllib.utils import try_import_tf
+
+tf = try_import_tf()
+
+
+@DeveloperAPI
+def add_time_dimension(padded_inputs, seq_lens):
+    """Adds a time dimension to padded inputs.
+
+    Arguments:
+        padded_inputs (Tensor): a padded batch of sequences. That is,
+            for seq_lens=[1, 2, 2], then inputs=[A, *, B, B, C, C], where
+            A, B, C are sequence elements and * denotes padding.
+        seq_lens (Tensor): the sequence lengths within the input batch,
+            suitable for passing to tf.nn.dynamic_rnn().
+
+    Returns:
+        Reshaped tensor of shape [NUM_SEQUENCES, MAX_SEQ_LEN, ...].
+    """
+
+    # Sequence lengths have to be specified for LSTM batch inputs. The
+    # input batch must be padded to the max seq length given here. That is,
+    # batch_size == len(seq_lens) * max(seq_lens)
+    padded_batch_size = tf.shape(padded_inputs)[0]
+    max_seq_len = padded_batch_size // tf.shape(seq_lens)[0]
+
+    # Dynamically reshape the padded batch to introduce a time dimension.
+    new_batch_size = padded_batch_size // max_seq_len
+    new_shape = ([new_batch_size, max_seq_len] +
+                 padded_inputs.get_shape().as_list()[1:])
+    return tf.reshape(padded_inputs, new_shape)
 
 
 @DeveloperAPI
